refactor: silence 1 warning related to integer signs.

+++ src/2-ec/ecEd.h
@@ -189,10 +189,10 @@ enum {
     ecEd_BytesVLong = qPrivateUseBegin + 4,
 };
 
-#define cCurveEd(q,bits) (                              \
-        q==ecEd_BytesOpCtx ? ECED_OPCTX_SIZE(bits) :    \
-        q==ecEd_BytesXYTZ ? ECED_XYTZ_SIZE(bits) :      \
-        q==ecEd_BytesVLong ? VLONG_BITS_SIZE(bits) :    \
+#define cCurveEd(q,bits) (                                      \
+        q==ecEd_BytesOpCtx ? (IntPtr)ECED_OPCTX_SIZE(bits) :    \
+        q==ecEd_BytesXYTZ ? (IntPtr)ECED_XYTZ_SIZE(bits) :      \
+        q==ecEd_BytesVLong ? (IntPtr)VLONG_BITS_SIZE(bits) :    \
         0)
 
 #define xCurveEd(q,bits,pCrvDef) (              \
